Moved logic from boarClient.tsx to separate files
